refactor: portal 사용 컨벤션 맞게 수정 및 폴더 이름 변경
