+++ apps/web/src/components/Portal/index.ts
@@ -1,0 +1,2 @@
+export * from './AnimatePortal';
+export * from './Portal';
